ARUHA-51: Moved class
